+++ demo/src/app/pages/pages.module.ts
@@ -20,7 +20,9 @@ import { AdvancedExampleConfirmComponent } from './demo/components/advanced-exam
 import { AdvancedExampleServerComponent } from './demo/components/advanced-example-server.component';
 import { AdvancedExamplesTypesComponent } from './demo/components/advanced-example-types.component';
 import { AdvancedExamplesCustomEditorComponent } from './demo/components/advanced-example-custom-editor.component';
+import { AdvancedExampleFiltersComponent } from './demo/components/advanced-example-filters.component';
 import { CustomEditorComponent } from './demo/components/custom-editor.component';
+import { CustomRenderComponent } from './demo/components/custom-render.component';
 
 @NgModule({
   imports: [
@@ -31,7 +33,10 @@ import { CustomEditorComponent } from './demo/components/custom-editor.component
     Ng2SmartTableModule,
     routing,
   ],
-  entryComponents: [CustomEditorComponent],
+  entryComponents: [
+    CustomEditorComponent,
+    CustomRenderComponent,
+  ],
   declarations: [
     DemoComponent,
     DocumentationComponent,
@@ -45,8 +50,10 @@ import { CustomEditorComponent } from './demo/components/custom-editor.component
     AdvancedExampleServerComponent,
     AdvancedExamplesTypesComponent,
     AdvancedExamplesCustomEditorComponent,
+    AdvancedExampleFiltersComponent,
     ScrollPositionDirective,
     CustomEditorComponent,
+    CustomRenderComponent,
   ],
 })
 export class PagesModule {