+++ src/ng2-smart-table/components/filter/filter-types/default-filter.ts
@@ -1,0 +1,33 @@
+import { Component, Input, Output, EventEmitter, OnInit, OnDestroy } from '@angular/core';
+import { Subscription } from 'rxjs/Subscription';
+
+import { Column } from '../../../lib/data-set/column';
+
+export class DefaultFilter implements Filter, OnDestroy {
+
+  delay: number = 300;
+  changesSubscription: Subscription;
+  @Input() query: string;
+  @Input() inputClass: string;
+  @Input() column: Column;
+  @Output() filter = new EventEmitter<string>();
+
+  ngOnDestroy() {
+    if (this.changesSubscription)
+      this.changesSubscription.unsubscribe();
+  }
+
+  setFilter() {
+    this.filter.emit(this.query);
+  }
+}
+
+export interface Filter {
+
+  delay?: number;
+  changesSubscription?: Subscription;
+  query: string;
+  inputClass: string;
+  column: Column;
+  filter: EventEmitter<string>;
+}