+++ src/ng2-smart-table/components/filter/filter-types/index.ts
@@ -1,0 +1,4 @@
+export * from './input-filter.component';
+export * from './select-filter.component';
+export * from './checkbox-filter.component';
+export * from './completer-filter.component';