+++ src/ng2-smart-table/components/thead/cells/checkbox-select-all.component.ts
@@ -1,0 +1,16 @@
+import { Component, Input } from '@angular/core';
+
+import { Grid } from '../../../lib/grid';
+
+@Component({
+  selector: '[ng2-st-checkbox-select-all]',
+  template: `
+    <input type="checkbox" [ngModel]="isAllSelected">
+  `
+})
+export class CheckboxSelectAllComponent {
+
+  @Input() grid: Grid;
+  @Input() source: any;
+  @Input() isAllSelected: boolean;
+}