+++ src/ng2-smart-table/components/thead/cells/title/title.component.ts
@@ -1,40 +1,36 @@
-import { Component, Input, Output, EventEmitter } from '@angular/core';
+import { Component, Input, Output, EventEmitter, OnInit } from '@angular/core';
 
-import { DataSource } from '../../lib/data-source/data-source';
-import { Column } from '../../lib/data-set/column';
+import { DataSource } from '../../../../lib/data-source/data-source';
+import { Column } from '../../../../lib/data-set/column';
 
 @Component({
   selector: 'ng2-smart-table-title',
-  styleUrls: ['title.scss'],
+  styleUrls: ['./title.component.scss'],
   template: `
-    <a href="#"
-      *ngIf="column.isSortable"
-      (click)="_sort($event, column)" 
-      class="ng2-smart-sort-link sort"
-      [ngClass]="currentDirection">
+    <a href="#" *ngIf="column.isSortable"
+                (click)="_sort($event, column)" 
+                class="ng2-smart-sort-link sort"
+                [ngClass]="currentDirection">
       {{ column.title }}
     </a>
     <span class="ng2-smart-sort" *ngIf="!column.isSortable">{{ column.title }}</span>
   `
 })
-export class TitleComponent {
-
-  @Input() column: Column;
-  @Input() source: DataSource;
-
-  @Output() sort = new EventEmitter<any>();
+export class TitleComponent implements OnInit {
 
   currentDirection = '';
+  @Input() column: Column;
+  @Input() source: DataSource;
+  @Output() sort = new EventEmitter<any>();
 
-  ngOnInit(): void {
+  ngOnInit() {
     this.source.onChanged().subscribe((elements) => {
       let sortConf = this.source.getSort();
 
-      if (sortConf.length > 0 && sortConf[0]['field'] === this.column.id) {
+      if (sortConf.length > 0 && sortConf[0]['field'] === this.column.id)
         this.currentDirection = sortConf[0]['direction'];
-      } else {
+      else
         this.currentDirection = '';
-      }
 
       sortConf.forEach((fieldConf) => {
 
@@ -42,7 +38,8 @@ export class TitleComponent {
     });
   }
 
-  _sort(): boolean {
+  _sort(event: any) {
+    event.preventDefault();
     this.changeSortDirection();
     this.source.setSort([
       {
@@ -52,7 +49,6 @@ export class TitleComponent {
       }
     ]);
     this.sort.emit(null);
-    return false;
   }
 
   changeSortDirection(): string {