+++ src/ng2-smart-table/lib/data-set/column.ts
@@ -11,10 +11,11 @@ export class Column {
   public sortDirection: string = '';
   public defaultSortDirection: string = '';
   public editor: { type: string, config: any, component: any } = { type: '', config: {}, component: null };
+  public filter: { type: string, config: any } = { type: '', config: {} };
+  public renderComponent: any = null;
   compareFunction: Function;
   valuePrepareFunction: Function;
   filterFunction: Function;
-  cellRenderFunction: Function;
 
   constructor(public id: string, protected settings: any, protected dataSet: DataSet) {
     this.process();
@@ -32,12 +33,16 @@ export class Column {
     return this.filterFunction;
   }
 
-  public getCellRenderFunction(): Function {
-    return this.cellRenderFunction;
-  }
-
   public getConfig(): any {
     return this.editor.config;
+  }
+
+  getFilterType(): any {
+    return this.filter && this.filter.type;
+  }
+
+  getFilterConfig(): any {
+    return this.filter && this.filter.config;
   }
 
   protected process(): void {
@@ -45,6 +50,8 @@ export class Column {
     this.class = this.settings['class'];
     this.type = this.prepareType();
     this.editor = this.settings['editor'];
+    this.filter = this.settings['filter'];
+    this.renderComponent = this.settings['renderComponent'];
 
     this.isFilterable = typeof this.settings['filter'] === 'undefined' ? true : !!this.settings['filter'];
     this.defaultSortDirection = ['asc', 'desc'].indexOf(this.settings['sortDirection']) !== -1 ? this.settings['sortDirection'] : '';
@@ -55,7 +62,6 @@ export class Column {
     this.compareFunction = this.settings['compareFunction'];
     this.valuePrepareFunction = this.settings['valuePrepareFunction'];
     this.filterFunction = this.settings['filterFunction'];
-    this.cellRenderFunction = this.settings['cellRenderFunction'];
   }
 
   prepareType(): string {