+++ CHANGELOG.md
@@ -1,9 +1,26 @@
+<a name="0.6.0-0"></a>
+# 0.6.0-0 (2017-03-02)
+
+### Features
+
+* **core:** configuring column filters ([#185](https://github.com/akveo/ng2-smart-table/pull/185))
+* **core:** custom cell view component ([#193](https://github.com/akveo/ng2-smart-table/pull/193))
+* **core:** update to angular 2.4.9
+
+### Bug Fixes
+
+* **demo:** fix url redirect to gh-pages ([8a54dea141](https://github.com/akveo/ng2-smart-table/commit/8a54dea141041d8db19f42430db46d2a2c2dcae8))
+
+### Code Refactoring
+
+* **core:** grouping components by section ([#172](https://github.com/akveo/ng2-smart-table/pull/172))
+
 <a name="0.3.1"></a>
 # 0.3.1 (2016-09-19)
 
 ### Features
 
-* Update to angular 2.0
+* **core:** update to angular 2.0
 
 ### How to update
 
@@ -15,7 +32,7 @@
 
 ### Features
 
-* Update to angular rc.5
+* **core:** update to angular rc.5
 
 ### How to update